pyproject.toml
[tool.poetry]
name = "storage"
version = "0.1.0"
description = "Comparing cold storage solutions for 1000 ~1Mb JSONs"
authors = ["mrtolkien <gary.mialaret@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
riotwatcher = "^3.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pymongo = "^4.1.1"
diskcache = "^5.4.0"
black = {version = "^22.3.0", allow-prereleases = true}
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
